+++ NEWS.md
@@ -6,6 +6,8 @@
 * Added new features in copyProject. Now empty.input = NA now only keeps files in the root "input" folder and input files in the sub folders that are actually used in the StoX project. Similarly, empty.output = NA implies to only keep the output files from enabled processes which are set to write output data. Also, the argument clean.process was added to clear the "process" folder of all files except the project.json file and the projectSession if close = FALSE.
 * Separated the comparison between original and new output out into the new exported function compareModelData(), so that this function can be used to compare projects without running the projects.
 * Added unit tests for the new functionality in copyProject().
+* Added the argument ignore.class in compareProjectToStoredOutputFiles(), which can be used to ignore process output of specific classes when comparing original and new data.
+* Fixed bug in zipProject() where using utils::zip() failed due to zip::zip() syntax (zip::zip() will not be used, as the base zip is regarded more reliable).
 
 
 # RstoxFramework v4.2.0 (2025-02-04)